# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "## TMF API Reference: TMF656 - Service Test Management\n\n### Release\
    \ : 18.5 - December 2018\n\nService Test Management API goal is to provide the\
    \ ability to manage tests of provisioned Services. \n\n### Resource \n- ServiceTest\
    \ \n- ServiceTestSpecification\n\n### Operations\nService Test Management API\
    \ performs the following operations on the resources :\n- Retrieve an entity or\
    \ a collection of entities depending on filter criteria\n- Partial update of an\
    \ entity (including updating rules)\n- Create an entity (including default values\
    \ and creation rules)\n- Delete an entity (for administration purposes)\n- Manage\
    \ notification of events"
  version: "4.0.0"
  title: "Service Test Management"
host: "localhost:8080"
basePath: "/tmf-api"
...
schemes:
- "http"
consumes:
- "application/json;charset=utf-8"
produces:
- "application/json;charset=utf-8"
paths:
...
  /serviceTestSpecification:
    get:
      tags:
      - "serviceTestSpecification"
      summary: "List or find ServiceTestSpecification objects"
      description: "This operation list or find ServiceTestSpecification entities"
      operationId: "listServiceTestSpecification"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to be provided in response"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "Requested index for start of resources to be provided in response"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Requested number of resources to be provided in response"
        required: false
        type: "integer"
      responses:
        200:
          description: "Success"
          headers:
            X-Total-Count:
              type: "integer"
              description: "Total number of items matching criteria"
            X-Result-Count:
              type: "integer"
              description: "Actual number of items returned in the response body"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceTestSpecification"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceTestSpecification"
    post:
      tags:
      - "serviceTestSpecification"
      summary: "Creates a ServiceTestSpecification"
      description: "This operation creates a ServiceTestSpecification entity."
      operationId: "createServiceTestSpecification"
      parameters:
      - in: "body"
        name: "serviceTestSpecification"
        description: "The ServiceTestSpecification to be created"
        required: true
        schema:
          $ref: "#/definitions/ServiceTestSpecification_Create"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/ServiceTestSpecification"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceTestSpecification"
  /serviceTestSpecification/{id}:
    get:
      tags:
      - "serviceTestSpecification"
      summary: "Retrieves a ServiceTestSpecification by ID"
      description: "This operation retrieves a ServiceTestSpecification entity. Attribute\
        \ selection is enabled for all first level attributes."
      operationId: "retrieveServiceTestSpecification"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceTestSpecification"
        required: true
        type: "string"
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to provide in response"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/ServiceTestSpecification"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceTestSpecification"
    delete:
      tags:
      - "serviceTestSpecification"
      summary: "Deletes a ServiceTestSpecification"
      description: "This operation deletes a ServiceTestSpecification entity."
      operationId: "deleteServiceTestSpecification"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceTestSpecification"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceTestSpecification"
    patch:
      tags:
      - "serviceTestSpecification"
      summary: "Updates partially a ServiceTestSpecification"
      description: "This operation updates partially a ServiceTestSpecification entity."
      operationId: "patchServiceTestSpecification"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceTestSpecification"
        required: true
        type: "string"
      - in: "body"
        name: "serviceTestSpecification"
        description: "The ServiceTestSpecification to be updated"
        required: true
        schema:
          $ref: "#/definitions/ServiceTestSpecification_Update"
      responses:
        200:
          description: "Updated"
          schema:
            $ref: "#/definitions/ServiceTestSpecification"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceTestSpecification"
...
definitions:
...
  ServiceSpecificationRef:
    type: "object"
    required:
    - "href"
    - "id"
    properties:
      id:
        type: "string"
        description: "Unique identifier of the service specification"
      href:
        type: "string"
        description: "Reference of the serviceSpecification"
      name:
        type: "string"
        description: "Name of the requiredServiceSpecification"
      version:
        type: "string"
        description: "Service specification version"
      targetServiceSchema:
        description: "A target service schema reference (TargetServiceSchemaRef).\
          \ The reference object to the schema and type of target service which is\
          \ described by service specification."
        $ref: "#/definitions/TargetServiceSchema"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
      '@referredType':
        type: "string"
        description: "The actual type of the target instance when needed for disambiguation."
    description: "Service specification reference: ServiceSpecification(s) required\
      \ to realize a ProductSpecification."
    example:
      targetServiceSchema:
        '@baseType': "@baseType"
        '@type': "@type"
        '@schemaLocation': "@schemaLocation"
      '@referredType': "@referredType"
      '@baseType': "@baseType"
      '@type': "@type"
      name: "name"
      id: "id"
      href: "href"
      '@schemaLocation': "http://example.com/aeiou"
      version: "version"
...
  ServiceTestSpecification:
    type: "object"
    properties:
      id:
        type: "string"
        description: "Identifier of a service test specification."
      href:
        type: "string"
        description: "Hyperlink to access a service test specification."
      description:
        type: "string"
        description: "Description of a service test specification."
      name:
        type: "string"
        description: "Name of a service test specification."
      version:
        type: "string"
        description: "Version of a service test specification "
      relatedServiceSpecification:
        description: "The related service specification."
        $ref: "#/definitions/ServiceSpecificationRef"
      testMeasureDefinition:
        type: "array"
        description: "A list of definitions for the measurements for the test defined\
          \ by this specification"
        items:
          $ref: "#/definitions/TestMeasureDefinition"
      validFor:
        description: "The period of time for which this specification is valid"
        $ref: "#/definitions/TimePeriod"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    description: "The service test specification describes the service test in terms\
      \ of parameters to be configured and \nmeasures to be taken."
    example:
      '@baseType': "@baseType"
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      testMeasureDefinition:
      - metricName: "metricName"
        unitOfMeasure: "unitOfMeasure"
        validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        '@type': "@type"
        captureFrequency: "captureFrequency"
        metricHref: "metricHref"
        metricDescription: "metricDescription"
        '@baseType': "@baseType"
        valueType: "valueType"
        name: "name"
        captureMethod: "captureMethod"
        capturePeriod:
          amount: ""
          '@baseType': "@baseType"
          '@type': "@type"
          units: "units"
          '@schemaLocation': "http://example.com/aeiou"
        '@schemaLocation': "http://example.com/aeiou"
        thresholdRule:
        - consequence:
          - repeatAction: true
            '@baseType': "@baseType"
            validFor:
              startDateTime: "2000-01-23T04:56:07.000+00:00"
              endDateTime: "2000-01-23T04:56:07.000+00:00"
            prescribeAction: "prescribeAction"
            '@type': "@type"
            name: "name"
            description: "description"
            '@schemaLocation': "http://example.com/aeiou"
          - repeatAction: true
            '@baseType': "@baseType"
            validFor:
              startDateTime: "2000-01-23T04:56:07.000+00:00"
              endDateTime: "2000-01-23T04:56:07.000+00:00"
            prescribeAction: "prescribeAction"
            '@type': "@type"
            name: "name"
            description: "description"
            '@schemaLocation': "http://example.com/aeiou"
          conformanceComparatorLower: "conformanceComparatorLower"
          conformanceComparatorUpper: "conformanceComparatorUpper"
          '@type': "@type"
          description: "description"
          thresholdRuleSeverity: "thresholdRuleSeverity"
          '@baseType': "@baseType"
          tolerancePeriod:
            amount: ""
            '@baseType': "@baseType"
            '@type': "@type"
            units: "units"
            '@schemaLocation': "http://example.com/aeiou"
          numberOfAllowedCrossing: 0
          conformanceTargetUpper: "conformanceTargetUpper"
          name: "name"
          '@schemaLocation': "http://example.com/aeiou"
          conformanceTargetLower: "conformanceTargetLower"
        - consequence:
          - repeatAction: true
            '@baseType': "@baseType"
            validFor:
              startDateTime: "2000-01-23T04:56:07.000+00:00"
              endDateTime: "2000-01-23T04:56:07.000+00:00"
            prescribeAction: "prescribeAction"
            '@type': "@type"
            name: "name"
            description: "description"
            '@schemaLocation': "http://example.com/aeiou"
          - repeatAction: true
            '@baseType': "@baseType"
            validFor:
              startDateTime: "2000-01-23T04:56:07.000+00:00"
              endDateTime: "2000-01-23T04:56:07.000+00:00"
            prescribeAction: "prescribeAction"
            '@type': "@type"
            name: "name"
            description: "description"
            '@schemaLocation': "http://example.com/aeiou"
          conformanceComparatorLower: "conformanceComparatorLower"
          conformanceComparatorUpper: "conformanceComparatorUpper"
          '@type': "@type"
          description: "description"
          thresholdRuleSeverity: "thresholdRuleSeverity"
          '@baseType': "@baseType"
          tolerancePeriod:
            amount: ""
            '@baseType': "@baseType"
            '@type': "@type"
            units: "units"
            '@schemaLocation': "http://example.com/aeiou"
          numberOfAllowedCrossing: 0
          conformanceTargetUpper: "conformanceTargetUpper"
          name: "name"
          '@schemaLocation': "http://example.com/aeiou"
          conformanceTargetLower: "conformanceTargetLower"
      - metricName: "metricName"
        unitOfMeasure: "unitOfMeasure"
        validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        '@type': "@type"
        captureFrequency: "captureFrequency"
        metricHref: "metricHref"
        metricDescription: "metricDescription"
        '@baseType': "@baseType"
        valueType: "valueType"
        name: "name"
        captureMethod: "captureMethod"
        capturePeriod:
          amount: ""
          '@baseType': "@baseType"
          '@type': "@type"
          units: "units"
          '@schemaLocation': "http://example.com/aeiou"
        '@schemaLocation': "http://example.com/aeiou"
        thresholdRule:
        - consequence:
          - repeatAction: true
            '@baseType': "@baseType"
            validFor:
              startDateTime: "2000-01-23T04:56:07.000+00:00"
              endDateTime: "2000-01-23T04:56:07.000+00:00"
            prescribeAction: "prescribeAction"
            '@type': "@type"
            name: "name"
            description: "description"
            '@schemaLocation': "http://example.com/aeiou"
          - repeatAction: true
            '@baseType': "@baseType"
            validFor:
              startDateTime: "2000-01-23T04:56:07.000+00:00"
              endDateTime: "2000-01-23T04:56:07.000+00:00"
            prescribeAction: "prescribeAction"
            '@type': "@type"
            name: "name"
            description: "description"
            '@schemaLocation': "http://example.com/aeiou"
          conformanceComparatorLower: "conformanceComparatorLower"
          conformanceComparatorUpper: "conformanceComparatorUpper"
          '@type': "@type"
          description: "description"
          thresholdRuleSeverity: "thresholdRuleSeverity"
          '@baseType': "@baseType"
          tolerancePeriod:
            amount: ""
            '@baseType': "@baseType"
            '@type': "@type"
            units: "units"
            '@schemaLocation': "http://example.com/aeiou"
          numberOfAllowedCrossing: 0
          conformanceTargetUpper: "conformanceTargetUpper"
          name: "name"
          '@schemaLocation': "http://example.com/aeiou"
          conformanceTargetLower: "conformanceTargetLower"
        - consequence:
          - repeatAction: true
            '@baseType': "@baseType"
            validFor:
              startDateTime: "2000-01-23T04:56:07.000+00:00"
              endDateTime: "2000-01-23T04:56:07.000+00:00"
            prescribeAction: "prescribeAction"
            '@type': "@type"
            name: "name"
            description: "description"
            '@schemaLocation': "http://example.com/aeiou"
          - repeatAction: true
            '@baseType': "@baseType"
            validFor:
              startDateTime: "2000-01-23T04:56:07.000+00:00"
              endDateTime: "2000-01-23T04:56:07.000+00:00"
            prescribeAction: "prescribeAction"
            '@type': "@type"
            name: "name"
            description: "description"
            '@schemaLocation': "http://example.com/aeiou"
          conformanceComparatorLower: "conformanceComparatorLower"
          conformanceComparatorUpper: "conformanceComparatorUpper"
          '@type': "@type"
          description: "description"
          thresholdRuleSeverity: "thresholdRuleSeverity"
          '@baseType': "@baseType"
          tolerancePeriod:
            amount: ""
            '@baseType': "@baseType"
            '@type': "@type"
            units: "units"
            '@schemaLocation': "http://example.com/aeiou"
          numberOfAllowedCrossing: 0
          conformanceTargetUpper: "conformanceTargetUpper"
          name: "name"
          '@schemaLocation': "http://example.com/aeiou"
          conformanceTargetLower: "conformanceTargetLower"
      '@type': "@type"
      name: "name"
      description: "description"
      id: "id"
      href: "href"
      '@schemaLocation': "http://example.com/aeiou"
      version: "version"
      relatedServiceSpecification:
        targetServiceSchema:
          '@baseType': "@baseType"
          '@type': "@type"
          '@schemaLocation': "@schemaLocation"
        '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
        version: "version"
...
  TimePeriod:
    type: "object"
    properties:
      endDateTime:
        type: "string"
        format: "date-time"
        description: "End of the time period, using IETC-RFC-3339 format"
      startDateTime:
        type: "string"
        format: "date-time"
        description: "Start of the time period, using IETC-RFC-3339 format. If you\
          \ define a start, you must also define an end"
    description: "A period of time, either as a deadline (endDateTime only) a startDateTime\
      \ only, or both"
    example:
      startDateTime: "2000-01-23T04:56:07.000+00:00"
      endDateTime: "2000-01-23T04:56:07.000+00:00"
...
  Error:
    type: "object"
    required:
    - "code"
    - "reason"
    properties:
      code:
        type: "integer"
        description: "Application relevant detail, defined in the API or a common\
          \ list."
      reason:
        type: "string"
        description: "Explanation of the reason for the error which can be shown to\
          \ a client user."
      message:
        type: "string"
        description: "More details and corrective actions related to the error which\
          \ can be shown to a client user."
      status:
        type: "integer"
        description: "HTTP Error code extension"
      referenceError:
        type: "string"
        format: "uri"
        description: "URI of documentation describing the error."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class."
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name."
    description: "Used when an API throws an Error, typically with a HTTP error response-code\
      \ (3xx, 4xx, 5xx)"